Pool - UI bugs on Pool detail page
